Move pkg/workloads to pkg/renderers
