+++ pkg/model/azure/azure.go
@@ -10,19 +10,19 @@ import (
 	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/model"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
+	"github.com/Azure/radius/pkg/renderers/containerv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/cosmosdbmongov1alpha1"
+	"github.com/Azure/radius/pkg/renderers/cosmosdbsqlv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/dapr"
+	"github.com/Azure/radius/pkg/renderers/daprpubsubv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/daprstatestorev1alpha1"
+	"github.com/Azure/radius/pkg/renderers/inboundroute"
+	"github.com/Azure/radius/pkg/renderers/keyvaultv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/manualscale"
+	"github.com/Azure/radius/pkg/renderers/mongodbv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/redisv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/servicebusqueuev1alpha1"
 	"github.com/Azure/radius/pkg/workloads"
-	"github.com/Azure/radius/pkg/workloads/containerv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/cosmosdbmongov1alpha1"
-	"github.com/Azure/radius/pkg/workloads/cosmosdbsqlv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/dapr"
-	"github.com/Azure/radius/pkg/workloads/daprpubsubv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/daprstatestorev1alpha1"
-	"github.com/Azure/radius/pkg/workloads/inboundroute"
-	"github.com/Azure/radius/pkg/workloads/keyvaultv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/manualscale"
-	"github.com/Azure/radius/pkg/workloads/mongodbv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/redisv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/servicebusqueuev1alpha1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
