+++ pkg/model/kubernetes/kubernetes.go
@@ -10,15 +10,15 @@ import (
 	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/model"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
+	"github.com/Azure/radius/pkg/renderers/containerv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/dapr"
+	"github.com/Azure/radius/pkg/renderers/daprstatestorev1alpha1"
+	"github.com/Azure/radius/pkg/renderers/inboundroute"
+	"github.com/Azure/radius/pkg/renderers/manualscale"
+	"github.com/Azure/radius/pkg/renderers/mongodbv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/rabbitmqv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/redisv1alpha1"
 	"github.com/Azure/radius/pkg/workloads"
-	"github.com/Azure/radius/pkg/workloads/containerv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/dapr"
-	"github.com/Azure/radius/pkg/workloads/daprstatestorev1alpha1"
-	"github.com/Azure/radius/pkg/workloads/inboundroute"
-	"github.com/Azure/radius/pkg/workloads/manualscale"
-	"github.com/Azure/radius/pkg/workloads/mongodbv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/rabbitmqv1alpha1"
-	"github.com/Azure/radius/pkg/workloads/redisv1alpha1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
