+++ pkg/renderers/cosmosdbsqlv1alpha1/render_test.go
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package keyvaultv1alpha1
+package cosmosdbsqlv1alpha1
 
 import (
 	"context"
@@ -11,24 +11,14 @@ import (
 
 	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/model/components"
-	"github.com/Azure/radius/pkg/radlogger"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
+	"github.com/Azure/radius/pkg/renderers"
+	"github.com/Azure/radius/pkg/renderers/cosmosdbmongov1alpha1"
 	"github.com/Azure/radius/pkg/workloads"
-	"github.com/go-logr/logr"
 	"github.com/stretchr/testify/require"
 )
 
-func createContext(t *testing.T) context.Context {
-	logger, err := radlogger.NewTestLogger(t)
-	if err != nil {
-		t.Log("Unable to initialize logger")
-		return context.Background()
-	}
-	return logr.NewContext(context.Background(), logger)
-}
-
 func Test_Render_Managed_Success(t *testing.T) {
-	ctx := createContext(t)
 	renderer := Renderer{}
 
 	workload := workloads.InstantiatedWorkload{
@@ -44,23 +34,39 @@ func Test_Render_Managed_Success(t *testing.T) {
 		BindingValues: map[components.BindingKey]components.BindingState{},
 	}
 
-	resources, err := renderer.Render(ctx, workload)
+	resources, err := renderer.Render(context.Background(), workload)
 	require.NoError(t, err)
-
 	require.Len(t, resources, 1)
-	resource := resources[0]
 
-	require.Equal(t, outputresource.LocalIDKeyVault, resource.LocalID)
-	require.Equal(t, outputresource.KindAzureKeyVault, resource.Kind)
+	renderedResource := resources[0]
+
+	require.Equal(t, outputresource.LocalIDAzureCosmosDBSQL, renderedResource.LocalID)
+	require.Equal(t, outputresource.KindAzureCosmosDBSQL, renderedResource.Kind)
 
 	expected := map[string]string{
-		handlers.ManagedKey: "true",
+		handlers.ManagedKey:              "true",
+		handlers.CosmosDBAccountBaseName: "test-component",
+		handlers.CosmosDBDatabaseNameKey: "test-component",
 	}
-	require.Equal(t, expected, resource.Resource)
+	require.Equal(t, expected, renderedResource.Resource)
+}
+
+func TestInvalidComponentKindFailure(t *testing.T) {
+	renderer := Renderer{}
+
+	workload := workloads.InstantiatedWorkload{
+		Workload: components.GenericComponent{
+			Name: "test-component",
+			Kind: cosmosdbmongov1alpha1.Kind,
+		},
+	}
+
+	_, err := renderer.Render(context.Background(), workload)
+	require.Error(t, err)
+	require.Equal(t, "the component was expected to have kind 'azure.com/CosmosDBSQL@v1alpha1', instead it is 'azure.com/CosmosDBMongo@v1alpha1'", err.Error())
 }
 
 func Test_Render_Unmanaged_Success(t *testing.T) {
-	ctx := createContext(t)
 	renderer := Renderer{}
 
 	workload := workloads.InstantiatedWorkload{
@@ -70,31 +76,32 @@ func Test_Render_Unmanaged_Success(t *testing.T) {
 			Kind: Kind,
 			Name: "test-component",
 			Config: map[string]interface{}{
-				"resource": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.KeyVault/vaults/test-vault",
+				"resource": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.DocumentDB/databaseAccounts/test-account/sqlDatabases/test-database",
 			},
 		},
 		BindingValues: map[components.BindingKey]components.BindingState{},
 	}
 
-	resources, err := renderer.Render(ctx, workload)
+	resources, err := renderer.Render(context.Background(), workload)
 	require.NoError(t, err)
 
 	require.Len(t, resources, 1)
 	resource := resources[0]
 
-	require.Equal(t, outputresource.LocalIDKeyVault, resource.LocalID)
-	require.Equal(t, outputresource.KindAzureKeyVault, resource.Kind)
+	require.Equal(t, outputresource.LocalIDAzureCosmosDBSQL, resource.LocalID)
+	require.Equal(t, outputresource.KindAzureCosmosDBSQL, resource.Kind)
 
 	expected := map[string]string{
-		handlers.ManagedKey:      "false",
-		handlers.KeyVaultIDKey:   "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.KeyVault/vaults/test-vault",
-		handlers.KeyVaultNameKey: "test-vault",
+		handlers.ManagedKey:              "false",
+		handlers.CosmosDBAccountIDKey:    "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.DocumentDB/databaseAccounts/test-account",
+		handlers.CosmosDBAccountNameKey:  "test-account",
+		handlers.CosmosDBDatabaseIDKey:   "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.DocumentDB/databaseAccounts/test-account/sqlDatabases/test-database",
+		handlers.CosmosDBDatabaseNameKey: "test-database",
 	}
 	require.Equal(t, expected, resource.Resource)
 }
 
 func Test_Render_Unmanaged_MissingResource(t *testing.T) {
-	ctx := createContext(t)
 	renderer := Renderer{}
 
 	workload := workloads.InstantiatedWorkload{
@@ -111,13 +118,12 @@ func Test_Render_Unmanaged_MissingResource(t *testing.T) {
 		BindingValues: map[components.BindingKey]components.BindingState{},
 	}
 
-	_, err := renderer.Render(ctx, workload)
+	_, err := renderer.Render(context.Background(), workload)
 	require.Error(t, err)
-	require.Equal(t, workloads.ErrResourceMissingForUnmanagedResource.Error(), err.Error())
+	require.Equal(t, renderers.ErrResourceMissingForUnmanagedResource.Error(), err.Error())
 }
 
 func Test_Render_Unmanaged_InvalidResourceType(t *testing.T) {
-	ctx := createContext(t)
 	renderer := Renderer{}
 
 	workload := workloads.InstantiatedWorkload{
@@ -127,13 +133,13 @@ func Test_Render_Unmanaged_InvalidResourceType(t *testing.T) {
 			Kind: Kind,
 			Name: "test-component",
 			Config: map[string]interface{}{
-				"resource": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.SomethingElse/vaults/test-vault",
+				"resource": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.SomethingElse/databaseAccounts/sqlDatabases/test-database",
 			},
 		},
 		BindingValues: map[components.BindingKey]components.BindingState{},
 	}
 
-	_, err := renderer.Render(ctx, workload)
+	_, err := renderer.Render(context.Background(), workload)
 	require.Error(t, err)
-	require.Equal(t, "the 'resource' field must refer to a KeyVault", err.Error())
+	require.Equal(t, "the 'resource' field must refer to a CosmosDB SQL Database", err.Error())
 }
