+++ pkg/renderers/dapr/decorator.go
