+++ pkg/renderers/daprpubsubv1alpha1/render.go
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/model/components"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
+	"github.com/Azure/radius/pkg/renderers"
 	"github.com/Azure/radius/pkg/workloads"
 )
 
@@ -71,7 +72,7 @@ func (r Renderer) Render(ctx context.Context, w workloads.InstantiatedWorkload) 
 		}
 
 		if component.Config.Resource != "" {
-			return nil, workloads.ErrResourceSpecifiedForManagedResource
+			return nil, renderers.ErrResourceSpecifiedForManagedResource
 		}
 
 		// generate data we can use to manage a servicebus topic
@@ -97,10 +98,10 @@ func (r Renderer) Render(ctx context.Context, w workloads.InstantiatedWorkload) 
 		}
 
 		if component.Config.Resource == "" {
-			return nil, workloads.ErrResourceMissingForUnmanagedResource
+			return nil, renderers.ErrResourceMissingForUnmanagedResource
 		}
 
-		topicID, err := workloads.ValidateResourceID(component.Config.Resource, TopicResourceType, "ServiceBus Topic")
+		topicID, err := renderers.ValidateResourceID(component.Config.Resource, TopicResourceType, "ServiceBus Topic")
 		if err != nil {
 			return nil, err
 		}
