+++ pkg/renderers/manualscale/decorator.go
