+++ pkg/renderers/manualscale/decorator_test.go
