+++ pkg/renderers/mongodbv1alpha1/azurerenderer.go
@@ -14,8 +14,8 @@ import (
 	"github.com/Azure/radius/pkg/model/components"
 	"github.com/Azure/radius/pkg/radlogger"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
+	"github.com/Azure/radius/pkg/renderers/cosmosdbmongov1alpha1"
 	"github.com/Azure/radius/pkg/workloads"
-	"github.com/Azure/radius/pkg/workloads/cosmosdbmongov1alpha1"
 )
 
 type AzureRenderer struct {
