+++ pkg/renderers/redisv1alpha1/azurerender_test.go
