+++ pkg/renderers/redisv1alpha1/kubernetesrender.go
