+++ pkg/renderers/redisv1alpha1/kubernetesrender_test.go
