+++ pkg/renderers/servicebusqueuev1alpha1/renderer.go
@@ -15,6 +15,7 @@ import (
 	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/model/components"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
+	"github.com/Azure/radius/pkg/renderers"
 	"github.com/Azure/radius/pkg/workloads"
 )
 
@@ -71,7 +72,7 @@ func (r Renderer) Render(ctx context.Context, w workloads.InstantiatedWorkload) 
 		}
 
 		if component.Config.Resource != "" {
-			return nil, workloads.ErrResourceSpecifiedForManagedResource
+			return nil, renderers.ErrResourceSpecifiedForManagedResource
 		}
 
 		// generate data we can use to manage a servicebus queue
@@ -91,10 +92,10 @@ func (r Renderer) Render(ctx context.Context, w workloads.InstantiatedWorkload) 
 		return []outputresource.OutputResource{resource}, nil
 	} else {
 		if component.Config.Resource == "" {
-			return nil, workloads.ErrResourceMissingForUnmanagedResource
+			return nil, renderers.ErrResourceMissingForUnmanagedResource
 		}
 
-		queueID, err := workloads.ValidateResourceID(component.Config.Resource, QueueResourceType, "ServiceBus Queue")
+		queueID, err := renderers.ValidateResourceID(component.Config.Resource, QueueResourceType, "ServiceBus Queue")
 		if err != nil {
 			return nil, err
 		}
